third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/slice.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] slice float32 1D constant tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test slice float32 actual 0 should be close enough to expected 0.5424534678459167 by ULP distance: expected a number less than or equal to 0n but got 1057676859n
[FAIL] [required] slice float16 1D constant tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test slice float16 actual 0 should be close enough to expected 0.54248046875 by ULP distance: expected a number less than or equal to 0 but got 14423
Harness: the test ran to completion.
